Count Good Substrings - 451D - Codeforces
http://codeforces.com/problemset/problem/451/D

A good question in which the major observation was that all
substrings starting with the same character are good substrings.
As soon as you understand that, the question is a simple counting
problem.

Also, removed the aquestionaday file because git commits seem
like a better place to do this anyways. From now on, there's
gonna be one commit per question.

PS: I need to get back to my rhythm, it took me a long time to solve
this and a pretty bad implementation bug should never have happened.