Karen and Game: 815A - Codeforces
http://codeforces.com/contest/815/problem/A

The idea is that you have n + m variables each representing the
number of times you apply the op on each row and column.
If we know the value of one of these variables, we can calculate
the values of the rest of them. So a simple linear search on
these values gives us the answer.

PS: Got two wrong answers - one because I was printing `column`
instead of `col` and another because I had written `i` instead
of `1` in an array subscript. This does not bode well for my
skills.