And Graph - 987F - Codeforces
https://codeforces.com/contest/987/problem/F

This was a problem with a really nice approach. Couldn't come
up with it myself, had to look at the editorial. Adding a number
of nodes to the graph in order to avoid having to calculate edges
is a nice concept though, I could see it in use in other
problems later.