Cut 'em all - 982C - Codeforces
Seemed like a dynamic programming problem, but you keep
making observations and it turns into just a simple dfs problem.

The greedy approach here is based on the idea that if you
can cut an edge (which is only possible if the size of the subtree
is even), you should cut it asap.

I made some thinking mistakes, making it more complicated than
it was, leading to a few wrong answers before I finally got it.